refactor: Structure to let grow and create github actions
