+++ package.json
@@ -2,15 +2,15 @@
   "name": "envilder",
   "version": "0.5.6",
   "description": "A CLI that securely centralizes your environment variables from AWS SSM as a single source of truth",
-  "main": "./lib/Cli.js",
+  "main": "./lib/apps/cli/Cli.js",
   "bin": {
-    "envilder": "lib/Cli.js"
+    "envilder": "lib/apps/cli/Cli.js"
   },
   "scripts": {
     "clean": "npm cache clean --force && npx rimraf lib && npx rimraf coverage && npx rimraf node_modules",
     "build": "tsc",
     "local:install": "npm run build && node --loader ts-node/esm scripts/pack-and-install.ts",
-    "local:test-run": "npm run build && node lib/cli/cli.js --map=tests/cli/sample/param-map.json --envfile=tests/cli/sample/autogenerated.env",
+    "local:test-run": "npm run build && node lib/apps/cli/Cli.js --map=tests/sample/param-map.json --envfile=tests/sample/autogenerated.env",
     "format": "biome format",
     "format:write": "biome format --write",
     "lint": "secretlint \"**/*\" && biome check --write && tsc --noEmit",
