+++ tests/envilder/application/dispatch/DispatchActionCommandHandler.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { DispatchActionCommand } from '../../../../src/cli/application/dispatch/DispatchActionCommand';
-import { DispatchActionCommandHandler } from '../../../../src/cli/application/dispatch/DispatchActionCommandHandler';
-import type { ExportSsmToEnvCommandHandler } from '../../../../src/cli/application/exportSsmToEnv/ExportSsmToEnvCommandHandler';
-import type { ImportEnvToSsmCommandHandler } from '../../../../src/cli/application/importEnvToSsm/ImportEnvToSsmCommandHandler';
-import type { PushSingleVariableCommandHandler } from '../../../../src/cli/application/pushSingleVariable/PushSingleVariableCommandHandler';
-import { OperationMode } from '../../../../src/cli/domain/OperationMode';
+import { DispatchActionCommand } from '../../../../src/envilder/application/dispatch/DispatchActionCommand';
+import { DispatchActionCommandHandler } from '../../../../src/envilder/application/dispatch/DispatchActionCommandHandler';
+import type { ExportSsmToEnvCommandHandler } from '../../../../src/envilder/application/exportSsmToEnv/ExportSsmToEnvCommandHandler';
+import type { ImportEnvToSsmCommandHandler } from '../../../../src/envilder/application/importEnvToSsm/ImportEnvToSsmCommandHandler';
+import type { PushSingleVariableCommandHandler } from '../../../../src/envilder/application/pushSingleVariable/PushSingleVariableCommandHandler';
+import { OperationMode } from '../../../../src/envilder/domain/OperationMode';
 
 const mockExportSsmToEnvCommandHandler = {
   handle: vi.fn(),
