+++ tests/envilder/application/exportSsmToEnv/ExportSsmToEnvCommandHandler.test.ts
@@ -8,11 +8,11 @@ import {
   type Mock,
   vi,
 } from 'vitest';
-import { ExportSsmToEnvCommand } from '../../../../src/cli/application/exportSsmToEnv/ExportSsmToEnvCommand';
-import { ExportSsmToEnvCommandHandler } from '../../../../src/cli/application/exportSsmToEnv/ExportSsmToEnvCommandHandler';
-import type { IEnvFileManager } from '../../../../src/cli/domain/ports/IEnvFileManager';
-import type { ILogger } from '../../../../src/cli/domain/ports/ILogger';
-import type { ISecretProvider } from '../../../../src/cli/domain/ports/ISecretProvider';
+import { ExportSsmToEnvCommand } from '../../../../src/envilder/application/exportSsmToEnv/ExportSsmToEnvCommand';
+import { ExportSsmToEnvCommandHandler } from '../../../../src/envilder/application/exportSsmToEnv/ExportSsmToEnvCommandHandler';
+import type { IEnvFileManager } from '../../../../src/envilder/domain/ports/IEnvFileManager';
+import type { ILogger } from '../../../../src/envilder/domain/ports/ILogger';
+import type { ISecretProvider } from '../../../../src/envilder/domain/ports/ISecretProvider';
 
 const testValues: Record<string, string> = {
   '/path/to/ssm/email': 'mockedEmail@example.com',
