+++ tests/envilder/infrastructure/logger/ConsoleLogger.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
-import { ConsoleLogger } from '../../../../src/cli/infrastructure/logger/ConsoleLogger';
+import { ConsoleLogger } from '../../../../src/envilder/infrastructure/Logger/ConsoleLogger';
+
 
 describe('ConsoleLogger', () => {
   const sut = new ConsoleLogger();
